Add libraries and build tools
The web app is to be built in Preact and Tailwind CSS. The `main` folder
contains the composition root of the web app, and the `test` folder
contains the unit test suites.

Vite is the build tool, providing native support for the JSX and PostCSS
transformations required by Preact and Tailwind CSS.

Yarn is the package manager. The project is configured with Yarn Berry,
which stores a cache of most third-party dependencies in the
`.yarn/cache` folder in the Git repository, effectively replacing the
usual `node_modules` folder.

Jest is the unit testing framework. It needs a so-called 'transformer'
in order to support TypeScript. The `ts-jest` transformer performs type
checking of the unit tests, supports the Preact dialect of JSX, produces
correct source maps, and works with the IntelliJ IDEA debugger.

ESLint is the linting tool. A range of plugins enable analysis of
TypeScript sources, JSX expressions, React-like applications, and test
suites in Jest. This commit introduces a stricter set of rules than the
recommended ones. If a rule turns out to be too strict, it can be
relaxed later on.