Introduce the profile content
This commit makes a responsive page design of the profile content. The
Perseverance rover acts as a stand-in profile, providing dummy
information.

The dummy profile picture is a self-portrait of the Perseverance rover.
Courtesy NASA/JPL-Caltech. See the JPL Image Use Policy:
https://www.jpl.nasa.gov/jpl-image-use-policy

This commit also adds `date-fns` as a runtime dependency and
`ts-essentials` as a development dependency.