Eliminate the flash of unstyled content
Upon the introduction of a pre-rendered HTML page, a so-called flash of
unstyled content may occur, during which the web browser has rendered
the page before loading and applying its stylesheet.

This commit applies a simple fix to this by hiding all content (except
the background colour) until the stylesheet has been loaded.