Redesign the profile page
This commit overhauls the design of the profile page:
- It introduces a list of quick facts (age, residence, academic
discipline, and work experience).
- It introduces a biography.
- It divides work experience and education into two separate timelines.
- It displays the month of an occupation period in addition to the year.
- It improves the utilisation of space on large screens, eliminating the
need for hyphenation.
- It has explicit support for four screen sizes and redefines the
breakpoint positions in `tailwind.config.cjs`.
- It applies brighter colours in default mode. Dark mode is to be
supported later.
- It displays a page footer.

Furthermore, this commit enables a new ESLint rule and adjusts an
existing one:
- react/no-multi-comp has been enabled. It disallows multiple component
declarations in a single file.
- @typescript-eslint/no-extra-parens accepts parentheses around
single-line JSX expressions in addition to multi-line ones.