Add an ESLint plugin for Tailwind CSS
The ESLint plugin has had support for Tailwind CSS 3 since December 27,
2021. It may discover class lists in JSX class attributes and within the
arguments to the `clsx()` function. However, it does not yet support
string interpolation of class attributes in JSX. For this reason, the
`clsx` function replaces string interpolation. The `ClassValue` type
from the `clsx` dependency replaces the custom-made `ClassList` type.

This commit updates ESLint to 8.7.0 and Tailwind CSS to 3.0.15.

It also rearranges the ESLint rules so that they are grouped by plugin
prefix (except for the `@typescript-eslint` rules that override built-in
ESLint rules). It opts in to the `tailwindcss/no-custom-classname` rule
to discover misspelled or deprecated classes. It opts in to the
`tailwindcss/migration-from-tailwind-2` rule to discover copy-pasted
code snippets that use Tailwind CSS 2 classes.