Rename language to locale
For now, locales remain represented by the `da` and `en` language codes.
While `da-DK` and `en-GB` would be more precise representations for
locales, the app does not support different language dialects, making
the detailed locale representation unnecessary.

In this commit, the app will now restore the selected locale when the
page is refreshed. The implementation is similar to that of theme
selections. `ThemeStoragePort` has been renamed to `ThemeCachePort` to
align its name with the new `LocaleCachePort` (as `LocaleStoragePort`
would easily be confused with the general concept of 'local storage').

`Language.test.ts` has been deleted as it is sufficiently covered by
`LocalePicker.test.tsx`.

All fake implementations of dependencies in unit tests have been moved
to a single folder `fakes` with its own path alias `+test/fakes`.

The `max-len` ESLint rule now allows a function declaration to exceed
the character limit when it does not declare any parameters.