Revert to complete client-side rendering
Currently, the server-oriented build produces an `index.html` page in
Danish with the light theme. However, these assumptions are broken when
the browser prefers the English locale or the dark theme.

Relying on these assumptions, the `hydrate` call in `main-client.tsx`
ends up rendering an incorrect DOM tree, most notably observed in the
initial appearance of the theme picker and the locale picker.

This commit reverts to calling `render` in `main-client.tsx`. It may
cause the browser to render the page in Danish for a brief moment before
Preact re-renders it in English.

Pre-rendering `index.html` is still desired to allow search engines to
discover the contents of the app, albeit only in Danish for now.