Extract button components
Refs and ARIA attributes must be applied to the final DOM elements in
the component hierarchy and should be forwarded through Preact
components via props.

The `overflow-x-hidden` class was not necessary in `SplitContainer`
after all, as the maximum size of the container would already cause
tables to overflow properly. However, the allocated two thirds of the
width did not take into account the horizontal gap between the
complementary section and the main section, allowing the content to grow
out of bounds.