Make the listbox component accessible by keyboard
The highlighted option in a listbox has been extracted into an explicit
piece of state instead of relying on CSS. This allows the keyboard event
handlers to change which option is currently highlighted.

The usage site of a listbox no longer has to explicitly forward the
configuration and state of the listbox button and options. This
information is resolved automatically through the surrounding Preact
context.

The checkmark icon of the currently selected listbox option is
automatically inserted by the listbox. `ListboxOption` has been renamed
to `LinearListboxOption` to emphasise that it has become a horizontal
flex container.

A `ReadonlyNonEmptyArray` type has been introduced to indicate arrays
that must contain at least one item. This allows the keyboard event
handlers of the listbox to skip checking that the length of the
`options` array is greater than 0.