Load images progressively
A progressive image is one that displays a low-quality placeholder while
loading the high-resolution image. The placeholder is sufficiently small
(less than 2 KB, as defined in `vite.config.ts`) to be embedded directly
into the build artifacts without requiring a separate HTTP request.

A responsive image is one that adapts its resolution to fit the device
or provides 'art direction' (e.g. cropping the image to fit a small
viewport).

The responsive breakpoints have been slightly adjusted to accommodate
the sizes of the responsive images.